% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editRatio.R
\name{editRatio}
\alias{editRatio}
\title{Calculate the edit distance ratio for phonetic transcriptions}
\usage{
editRatio(x)
}
\arguments{
\item{x}{a data frame that contains target transcription and produced transcription on alternating rows}
}
\value{
The edit distance ratio, the number of edits (Levenshtein, 1966) divided by the number of target or produced segments, whichever is greater.
}
\description{
\code{editRatio} returns the edit distance ratio comparing the target transcription(s) to the produced transcription(s).
}
\section{References}{

Levenshtein, V. I. (1966). Binary codes capable of correcting deletions, insertions and reversals. Soviet Physics Doklady, 10 (8).
}

\examples{
x<- data.frame("Transcription"=c("Saul", "Paul"))
editRatio(x)

}
